auditia-api-python/requirements.txt
fastapi
resend
stripe
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
sh
passlib[bcrypt]
python-jose[cryptography]
python-multipart
